move files to schema
